housekeeping, add eslint, prettier
