+++ src/App.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import IssTracker from './components/IssTracker'
+
+const App = props => (
+  <div>
+    <IssTracker />
+  </div>
+)
+
+export default App
